Some changes

+++ RobotsOnMars/Application/View.h
@@ -10,10 +10,10 @@ class View
 {
 public:
     View() : offsetX{0}, offsetY{0} {}
-    //char *color(int color, char *Message);
+    //char *color(size_t color, char *Message);
     void Render(Robot *robot, GameArea &map);
 
 private:
-    int offsetX;
-    int offsetY;
+    size_t offsetX;
+    size_t offsetY;
 };
